requirements.txt
# =======================================================================================
# requirements.txt - Python Dependencies
# =======================================================================================

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pyserial==3.5
python-multipart==0.0.6
pydantic==2.5.0
